refactoring files. Comments
