+++ examples/replay/__init__.py
@@ -1,0 +1,38 @@
+import os
+import shutil
+import stat
+
+
+def copytree(src, dst, symlinks=False, ignore=None):
+    if not os.path.exists(dst):
+        os.makedirs(dst)
+        shutil.copystat(src, dst)
+    lst = os.listdir(src)
+    if ignore:
+        excl = ignore(src, lst)
+        lst = [x for x in lst if x not in excl]
+    for item in lst:
+        s = os.path.join(src, item)
+        d = os.path.join(dst, item)
+        if symlinks and os.path.islink(s):
+            if os.path.lexists(d):
+                os.remove(d)
+            os.symlink(os.readlink(s), d)
+            try:
+                st = os.lstat(s)
+                mode = stat.S_IMODE(st.st_mode)
+                os.lchmod(d, mode)
+            except:
+                pass  # lchmod not available
+        elif os.path.isdir(s):
+            copytree(s, d, symlinks, ignore)
+        else:
+            shutil.copy2(s, d)
+
+
+def copy_scenarios(save_dir, scenarios):
+    for i in range(len(scenarios)):
+        new_scenario_location = os.path.join(save_dir, scenarios[i])
+        if not os.path.exists(new_scenario_location):
+            copytree(scenarios[i], new_scenario_location)
+        scenarios[i] = new_scenario_location
