+++ zoo/policies/replay_agent.py
@@ -11,7 +11,8 @@ agent_index = 0
 
 
 class ReplayAgent(Agent):
-    """A helper agent that wraps another agent to allow replay of the agent actions"""
+    """A helper agent that wraps another agent to allow replay of the agent inputs and actions
+    Checkout bin/crash_test/README.md on how to use it"""
 
     def __init__(self, save_directory, id, read: bool, internal_spec: AgentSpec):
         import smarts.core
